fix(desktop): move logs toggle button from status bar to left sidebar
