+++ packages/desktop/src/renderer/components/LeftRail.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Settings, HelpCircle, MessageSquare } from 'lucide-react';
 import { cn } from '../lib/utils';
-import { usePluginLayoutStore, useSettingsStore } from '../store';
+import { usePluginLayoutStore, useSettingsStore, useUIStore } from '../store';
 import { PluginIcon } from './plugins/PluginIcon';
 
 interface RailButtonProps {
@@ -31,6 +31,8 @@ function RailButton({ active, onClick, title, children }: RailButtonProps): Reac
 export function LeftRail(): React.ReactElement {
   const contributions = usePluginLayoutStore((s) => s.contributions).filter((c) => c.zone === 'left-panel');
   const activeLeftPanelId = usePluginLayoutStore((s) => s.activeLeftPanelId);
+  const panelVisible = useUIStore((s) => s.panelVisible);
+  const setPanelVisible = useUIStore((s) => s.setPanelVisible);
 
   const handleSessionsClick = (): void => {
     usePluginLayoutStore.getState().setActiveLeftPanel(null);
@@ -75,6 +77,21 @@ export function LeftRail(): React.ReactElement {
         <RailButton onClick={() => useSettingsStore.getState().open(undefined, 'about')} title="Help">
           <HelpCircle size={16} />
         </RailButton>
+
+        {/* Logs toggle button */}
+        <div className="flex items-center justify-center gap-1 text-xs text-mf-text-secondary px-2">
+          <div className="text-mf-divider">|</div>
+          <button
+            onClick={() => setPanelVisible(!panelVisible)}
+            className={cn(
+              'px-2 py-1 rounded transition-colors',
+              panelVisible ? 'text-mf-text-primary' : 'text-mf-text-secondary hover:text-mf-text-primary',
+            )}
+            title="Toggle logs panel"
+          >
+            Logs
+          </button>
+        </div>
       </div>
     </div>
   );
